Schematic/board v2: add CH32V003 to read board ID and control power/sleep
The microcontroller allows us to actually monitor the BOARD_ID and USB CC1/2
lines to monitor and control power consumption.  It uses the unused 3rd
USB port on the hub.  ESD protection has also been added to all touchable
connectors.  A programming connector for the CH32V003 has been added based
on the same molex 2mm-pitch spring connector we use for input modules.

CH32V003 USB interface based on
  https://github.com/cnlohr/rv003usb
and we expect to flash the USB bootloader on this to allow it to be
reprogrammed by the end user:
  https://github.com/cnlohr/rv003usb/tree/master/bootloader

FreeCAD model updated.  Switched to realthunder FreeCAD fork since
it handled the parametric dependency updates better.