Tweak bottom clearances, change to slot for magnets on PCB
The bottom clearances weren't quite right around the pogo pin area, and
although my own PCBs made appropriate contact with the pogo pins, the
official Framework LED Matrix input module was about 1mm (!) short of
proper contact with the pins/magnet.  A couple of things:
* My "5mm diameter x 2mm thick" magnets were actually only 1.6mm thick
* It appears that we need an extra 0.5mm clearance between the "thick
  component" area and the pogo pin area.  The nominal clearance here is
  0.4mm but the Framework-supplied module was contacting here.
* The pogo pin plane should have been measured from the PCB bottom,
  which is 0.5mm above the "spacer bottom" level.

That last adjustment also moved the USB connector up by proxy, so we
needed to add a little bit of a bump over it to ensure a reasonable
minimum width for the shell here.  It's possible we'll remove the top
entirely here and fold this into the lid, but that complicates printing
the lid on a FDM 3d printer because the lid top is no longer planer. So
leave it as-is with a bit of a bump for now.

I also adjusted the cutout in the case so that it would accomodate the
original PCB design as well as the new one, since I have 5 perfectly
functional v1 boards.

The magnets also weren't very grabby.  This is party because they were
too far away from the steel (see above) but I decided to beef this up as
well.  Looking at the FW16 midplate it appears that a pair of opposite
polarity magnets is used.  Switch to an oval slot on the PCB so that we
can use a pair of magnets in our version as well.